+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - UI
   - the "Skip" button on the initialization panel is now enabled after 10 seconds [#310](https://github.com/vinifmor/bauh/issues/310)
   - faster package icons download
+  - faster packages filtering (`type`, `category`, `name`, etc... up to **95% less time**)
 
 ### Fixes
 - AppImage